feat(root): update syntax and terraform provider for AWS (#15)
Please find in the following pull request an update to the Terraform
provider to match the requirements of the new WebACL Module. This update
ensures compatibility and allows deployment within the same Spacelift
stack. Minor syntax issues have been fixed to meet the Terraform lint
standards as per our pre-commit configuration. Due to the nature of
these changes, I have bumped the module version to 3.0.0.

---------

Signed-off-by: Jonatan Mata <Jonatan.Mata@mdlz.com>