Add missing aws_api_method_settings, which allows for access control logging (#27)
In the process of adding the aws_api_method_settings, I realized finally
the better way to handle multiple stages.

In addition to the method_settings, there is now mapping of the hard
coded word "main" to "prod". This will rename/handle anytime someone is
using var.spacelift_stack_branch as the value to feed the stage
definition. For this reason alone I made this a breaking change, however
I assume it won't cause problems beyond a destroy/create.

I also added testing for the "simple" use case which had been broken in
previous releases.

Added an issue to delegate a r53 subdomain to this account via aft so we
can run tests with custom domains.